+++ frontend/src/components/Sidebar/SidebarMenu/sidebarMenuLinks.ts
@@ -1,0 +1,39 @@
+export const sidebarMenuLinks = [
+  {
+    name: "Projects",
+    url: "/projects",
+  },
+  {
+    name: "Administration Forms",
+    items: [
+      {
+        name: "Contacts",
+        url: "/admin/contacts",
+      },
+      {
+        name: "Suppliers",
+        url: "/admin/suppliers",
+      },
+      {
+        name: "Subcontractors",
+        url: "/admin/subcontractors",
+      },
+      {
+        name: "Resources",
+        url: "/admin/resources",
+      },
+      {
+        name: "Ministries / Org Name",
+        url: "/admin/ministries",
+      },
+      {
+        name: "User Management",
+        url: "/admin/users",
+      },
+      {
+        name: "Glossary",
+        url: "/admin/glossary",
+      },
+    ],
+  },
+];