+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 * frontend
   * added react-query
 
+### May 16, 2022 DESCW-362
+* frontend
+  * Added sub menu for admin links
+
 ### May 6, 2022
 * DESCW-271 - improve linter standards
   * frontend and backend